requirements.txt
# Core
vanna[gemini,sqlite]>=2.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
plotly>=5.22.0
pandas>=2.2.0
python-dotenv>=1.0.1

# LLM: Google Gemini (Option A)
google-genai>=1.5.0

# Utilities
pydantic>=2.7.0